Major code cleanup, part 1
- Bootstrap cleanup coming soon

Begin bootstrapping code rewrite

- Also other minor things

Cleanup:

- Begin bootstrapping cleanup
- Split into seperate files
- Hugely clean fun.c
