+++ electra/app/view/CSGradientView.m
