+++ electra/bootstrap/jailbreakd.plist
