+++ electra/exploit/early_kalloc.c
