+++ electra/exploit/early_kalloc.h
