+++ electra/exploit/kmem.c
