+++ electra/exploit/kmem.h
