+++ electra/exploit/symbols.c
