+++ electra/exploit/symbols.h
