+++ electra/the fun part/unlocknvram.c
