+++ electra/the fun part/unlocknvram.h
