+++ electra/the fun part/utilities/utils.h
@@ -1,0 +1,16 @@
+//
+//  utils.h
+//  electra
+//
+//  Created by Jamie on 27/01/2018.
+//  Copyright © 2018 Ian Beer. All rights reserved.
+//
+
+#ifndef utils_h
+#define utils_h
+
+#include <stdio.h>
+
+//int system(const char *cmd);
+
+#endif /* utils_h */
